solve conflig merge review  into media details template
