+++ app/src/main/java/com/berlin/aflami/di/dataSourceModule.kt
@@ -5,14 +5,10 @@ import com.berlin.local.datasource.SearchLocalDataSourceImpl
 import com.berlin.remote.MovieDetailsRemoteDataSourceImpl
 import com.berlin.remote.SearchRemoteDataSourceImpl
 import com.berlin.remote.TvShowDetailsRemoteDataSourceImpl
-import com.berlin.remote.MovieDetailsRemoteDataSourceImpl
-import com.berlin.remote.SearchRemoteDataSourceImp
-import com.berlin.remote.SeriesDetailsRemoteDataSourceImpl
 import com.berlin.repository.datasource.local.SearchLocalDataSource
 import com.berlin.repository.datasource.remote.MovieDetailsRemoteDataSource
 import com.berlin.repository.datasource.remote.SearchRemoteDataSource
 import com.berlin.repository.datasource.remote.TvShowDetailsRemoteDataSource
-import com.berlin.repository.datasource.remote.SeriesDetailsRemoteDataSource
 import io.ktor.client.HttpClient
 import org.koin.dsl.module
