+++ app/src/main/java/com/berlin/aflami/di/repositoryModule.kt
@@ -2,15 +2,15 @@ package com.berlin.aflami.di
 
 import com.berlin.repository.MovieDetailsRepositoryImpl
 import com.berlin.repository.SearchRepositoryImpl
-import com.berlin.repository.SeriesDetailsRepositoryImpl
+import com.berlin.repository.TvShowDetailsRepositoryImpl
 import org.koin.dsl.module
 import repository.MovieDetailsRepository
 import repository.SearchRepository
-import repository.SeriesDetailsRepository
+import repository.TvShowDetailsRepository
 
 val repositoryModule = module {
     single<SearchRepository> { SearchRepositoryImpl(get(), get()) }
     single<MovieDetailsRepository> { MovieDetailsRepositoryImpl(get()) }
     single<TvShowDetailsRepository> { TvShowDetailsRepositoryImpl(get()) }
-    single<MovieDetailsRepository> { MovieDetailsRepositoryImpl(get()) }
 }
+
