+++ app/src/main/java/com/berlin/aflami/di/usecaseModule.kt
@@ -5,6 +5,7 @@ import usecase.GetMovieDetailsUseCase
 import usecase.GetMovieGalleryUseCase
 import usecase.SearchByActorNameUseCase
 import usecase.GetMovieCastUseCase
+import usecase.GetMovieReviewUseCase
 import usecase.GetSearchMoviesUseCase
 import usecase.GetSearchTvShowsUseCase
 import usecase.GetSimilarMoviesUseCase
@@ -12,6 +13,7 @@ import usecase.GetSimilarSeriesUseCase
 import usecase.GetTvShowDetailsUseCase
 import usecase.GetSeriesGalleryUseCase
 import usecase.GetSeriesCastUseCase
+import usecase.GetSeriesReviewUseCase
 import usecase.SearchByCountryUseCase
 
 val useCaseModule = module {
@@ -23,12 +25,10 @@ val useCaseModule = module {
     single { GetSimilarSeriesUseCase(get()) }
     single { GetMovieDetailsUseCase(get()) }
     single { GetTvShowDetailsUseCase(get()) }
-
     single { GetMovieGalleryUseCase(get()) }
     single { GetSeriesGalleryUseCase(get()) }
     single { GetMovieCastUseCase(get()) }
     single { GetSeriesCastUseCase(get()) }
-
     single { GetMovieReviewUseCase(get()) }
     single { GetSeriesReviewUseCase(get()) }
 }
