--- data/remote/src/main/kotlin/com/berlin/remote/SeriesDetailsRemoteDataSourceImpl.kt
@@ -1,15 +1,0 @@
-package com.berlin.remote
-
-import com.berlin.repository.datasource.remote.SeriesDetailsRemoteDataSource
-import com.berlin.repository.datasource.remote.dto.ReviewResponse
-import io.ktor.client.HttpClient
-import io.ktor.client.call.body
-import io.ktor.client.request.get
-
-class SeriesDetailsRemoteDataSourceImpl (
-    private val client: HttpClient
-): SeriesDetailsRemoteDataSource {
-    override suspend fun getReviews(id: Long): ReviewResponse {
-        return client.get(ApiConstants.SERIES_REVIEW
-            .replace("{series_id}", id.toString())).body()    }
-}
