+++ data/repository/src/main/java/com/berlin/repository/MovieDetailsRepositoryImpl.kt
@@ -6,6 +6,7 @@ import com.berlin.entity.Movie
 import com.berlin.repository.datasource.remote.MovieDetailsRemoteDataSource
 import com.berlin.repository.mapper.toDomain
 import com.berlin.entity.MovieDetails
+import com.berlin.entity.Review
 import exceptions.AflamiExceptions
 import com.berlin.repository.mapper.POSTER_PREFIX
 import repository.MovieDetailsRepository
