--- data/repository/src/main/java/com/berlin/repository/datasource/remote/SeriesDetailsRemoteDataSource.kt
@@ -1,8 +1,0 @@
-package com.berlin.repository.datasource.remote
-
-import com.berlin.repository.datasource.remote.dto.ReviewResponse
-
-interface SeriesDetailsRemoteDataSource {
-    suspend fun getReviews(id: Long): ReviewResponse
-
-}
