+++ data/repository/src/main/java/com/berlin/repository/datasource/remote/TvShowDetailsRemoteDataSource.kt
@@ -2,6 +2,7 @@ package com.berlin.repository.datasource.remote
 
 import com.berlin.repository.datasource.remote.dto.MediaCastResponse
 import com.berlin.repository.datasource.remote.dto.MediaImagesResponse
+import com.berlin.repository.datasource.remote.dto.ReviewResponse
 import com.berlin.repository.datasource.remote.dto.TVShowDetailsDto
 import com.berlin.repository.datasource.remote.dto.TVShowResponse
 
@@ -10,4 +11,5 @@ interface TvShowDetailsRemoteDataSource {
     suspend fun getTvShowDetails(seriesId: Long, language: String): TVShowDetailsDto
     suspend fun getSeriesCastDetails(seriesId: Long, language: String): MediaCastResponse
     suspend fun getSeriesSimilar(seriesId: Long): TVShowResponse
+    suspend fun getReviews(id: Long): ReviewResponse
 }
