--- domain/usecase/src/main/kotlin/repository/SeriesDetailsRepository.kt
@@ -1,8 +1,0 @@
-package repository
-
-import com.berlin.entity.Review
-
-interface SeriesDetailsRepository {
-    suspend fun getReviews(id: Long): List<Review>
-
-}
