+++ domain/usecase/src/main/kotlin/repository/TvShowDetailsRepository.kt
@@ -1,6 +1,7 @@
 package repository
 
 import com.berlin.entity.MediaCast
+import com.berlin.entity.Review
 import com.berlin.entity.TVShow
 import com.berlin.entity.TvShowDetails
 
@@ -9,5 +10,6 @@ interface TvShowDetailsRepository {
     suspend fun getSeriesImages(id:Long):List<String>
     suspend fun getSeriesCastDetails(seriesId: Long, language: String): List<MediaCast>
     suspend fun getSeriesSimilar(seriesId:Long):List<TVShow>
+    suspend fun getReviews(id: Long):List<Review>
 
 }
