+++ domain/usecase/src/main/kotlin/usecase/GetSeriesReviewUseCase.kt
@@ -1,10 +1,10 @@
 package usecase
 
 import com.berlin.entity.Review
-import repository.SeriesDetailsRepository
+import repository.TvShowDetailsRepository
 
 class GetSeriesReviewUseCase(
-    private val seriesDetailsRepository: SeriesDetailsRepository
+    private val tvShowDetailsRepository: TvShowDetailsRepository
 ) {
-    suspend operator fun invoke(id: Long): List<Review> = seriesDetailsRepository.getReviews(id)
+    suspend operator fun invoke(id: Long): List<Review> = tvShowDetailsRepository.getReviews(id)
 }
