+++ presentation/viewModel/src/main/kotlin/com/berlin/aflami/viewmodel/mediadetails/MediaType.kt
@@ -5,9 +5,6 @@ data class MovieDetailsTabsUiState(
     val tab: MovieDetailsTabs = MovieDetailsTabs.REVIEWS,
     val isSelected: Boolean = false,
 )
-enum class MediaType {
-    MOVIE, SERIES
-}
 
 enum class MovieDetailsTabs() {
     MORE_LIKE_THIS,
